Codebase Cleanup, Badger Query Optimization, PostgreSQL Index Creation

+++ sql/postgres/postgres.go
@@ -10,17 +10,27 @@ import (
 	_ "github.com/lib/pq" //postgres driver
 )
 
+var (
+	// RecreateTables indicates that you accept all consequences from dropping your table and recreating it
+	RecreateTables = "recreate tables"
+)
+
 // Options are the postgres datastore options, reexported here for convenience.
 type Options struct {
-	Host           string
-	Port           string
-	User           string
-	Password       string
-	Database       string
-	Table          string
-	SSLMode        string
-	RunMigrations  bool
-	RecreateTables bool
+	// AcceptRecreateWarning is used as a safety check to pevent accidental deletion of existing tables and data
+	// To accept the warning, you must set the value of this field to `recreate tables`, which can be done manually
+	// or via the usage of the public `RecreateTables` variable
+	AcceptRecreateWarning string
+	Host                  string
+	Port                  string
+	User                  string
+	Password              string
+	Database              string
+	Table                 string
+	SSLMode               string
+	RunMigrations         bool
+	RecreateTables        bool
+	CreateIndex           bool
 }
 
 // Queries are the postgres queries for a given table.
@@ -106,7 +116,8 @@ func (opts *Options) Create() (*sqlds.Datastore, error) {
 	if err != nil {
 		return nil, err
 	}
-	if opts.RecreateTables {
+	// only recreate the tables *IF* warning is accepted
+	if opts.RecreateTables && opts.AcceptRecreateWarning == RecreateTables {
 		if _, err := db.Exec(
 			"DROP TABLE IF EXISTS blocks",
 		); err != nil {
@@ -118,6 +129,17 @@ func (opts *Options) Create() (*sqlds.Datastore, error) {
 			"CREATE TABLE IF NOT EXISTS blocks (key TEXT NOT NULL UNIQUE, data BYTEA NOT NULL)",
 		); err != nil {
 			return nil, multierr.Combine(err, db.Close())
+		}
+		if opts.CreateIndex {
+			if _, err := db.Exec(
+				fmt.Sprintf(
+					"CREATE INDEX IF NOT EXISTS %s_key_text_pattern_ops_idx ON %s (key text_pattern_ops)",
+					opts.Table, opts.Table,
+				),
+			); err != nil {
+				return nil, multierr.Combine(err, db.Close())
+			}
+
 		}
 	}
 	return sqlds.NewDatastore(db, NewQueries(opts.Table)), nil
